+++ R/TempSourceDB.R
@@ -34,3 +34,41 @@
 # adduct_nl lib ----------------------------------------------------------------
 # load('~/Project/04_package/MetDNA2/data/lib_adduct_nl.rda')
 # usethis::use_data(lib_adduct_nl, overwrite = TRUE)
+#
+# ################################################################################
+# ms2_dodd_lib -----------------------------------------------------------------
+# load('~/Project/04_package/00_Database/DoddLib/02_ms2_lib/v230511_v2.1.0/ms2_dodd_lib_v2.1.0_230511.RData')
+# usethis::use_data(ms2_dodd_lib, overwrite = TRUE)
+#
+################################################################################
+# # ms2_dodd_lib -----------------------------------------------------------------
+# load('~/Project/04_package/00_Database/DoddLib/02_ms2_lib/v230612_v2.2.0/ms2_dodd_lib_v2.2.0_230612.RData')
+# usethis::use_data(ms2_dodd_lib, overwrite = TRUE)
+
+################################################################################
+# # ms2 peptide ----------------------------------------------------------------
+# load('~/Project/04_package/00_Database/FiehnPeptide/fiehn_peptide_lib_230727.RData')
+# usethis::use_data(fiehn_peptide_lib, overwrite = TRUE)
+
+################################################################################
+# # ms2_dodd_lib -----------------------------------------------------------------
+# load('~/Project/04_package/00_Database/DoddLib/02_ms2_lib/v231206_v2.3.0/ms2_dodd_lib_v2.3.0_231206.RData')
+# usethis::use_data(ms2_dodd_lib, overwrite = TRUE)
+
+# # cpd_dodd_lib -----------------------------------------------------------------
+# load('~/Project/04_package/00_Database/DoddLib/01_compound_RT_lib/04_version_231206/cpd_dodd_lib_v2.0.0_231206.RData')
+# usethis::use_data(cpd_dodd_lib, overwrite = TRUE)
+
+
+###############################################################################
+# load('~/Project/04_package/00_Database/GNPS_ReverseLib/00_converted_database/ms2_gnps_bile_acid_lib.RData')
+# gnps_bile_acid_lib <- ms2_gnps_bile_acid_lib
+# usethis::use_data(gnps_bile_acid_lib)
+#
+# load('~/Project/04_package/00_Database/GNPS_ReverseLib/00_converted_database/ms2_gnps_acyl_amides_lib.RData')
+# gnps_acyl_amides_lib <- ms2_gnps_acyl_amides_lib
+# usethis::use_data(gnps_acyl_amides_lib)
+#
+# load('~/Project/04_package/00_Database/GNPS_ReverseLib/00_converted_database/ms2_gnps_acyl_esters_lib.RData')
+# gnps_acyl_esters_lib <- ms2_gnps_acyl_esters_lib
+# usethis::use_data(gnps_acyl_esters_lib)