+++ NEWS.md
@@ -1,3 +1,27 @@
 #### DoddLabDatabase 0.1.0
 * Added a `NEWS.md` file to track changes to the package.
 * Initial version 0.1.0 (20230306)
+
+
+#### DoddLabDatabase 0.1.1
+* Added build DB function (retrieve_ms2, group_ms2, merge_ms2, select_best_ms2) to process STD ms2 data.
+
+#### DoddLabDatabase 0.1.2
+* update ms2_dodd_lib from v1.0.0 to v2.1.0
+
+#### DoddLabDatabase 0.1.3
+* update ms2_dodd_lib from v2.1.0 to v2.2.0
+
+#### DoddLabDatabase 0.1.4
+* Add fiehn_peptide_lib & modify related functions
+
+#### DoddLabDatabase 0.1.5
+* update ms2_dodd_lib from v2.2.0 to v2.3.0
+* update cpd_dodd_lib from v1.0.0 to v2.0.0
+
+#### DoddLabDatabase 0.2.0
+* Add gnps_bile_acid_lib, gnps_acyl_amides_lib, gnps_acyl_esters_lib
+* Add all_public option to load all public database one time
+
+#### DoddLabDatabase 0.2.1
+* Add adduct filter for public database